Fixing bugs in citizens transfer flow

+++ services/files/src/handlers/files.py
@@ -75,6 +75,9 @@ class Handler:
         try:
             path = os.path.join("/", self.bucket, user_id)
 
+            if not self.fs.isdir(path):
+                return make_response("No files found to delete", 201)
+
             self.fs.delete(path, recursive=True)
 
             return make_response("Deleted", 201)
